refactor: make button component to common and atoms folder
