+++ src/ui/atoms/CardFrame/index.ts
@@ -1,0 +1,1 @@
+export * from './CardFrame';
